ocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
anthropic
Pillow==11.0.0
python-dotenv==1.0.1
